Upgrade all projects to .Net 8
Some of the old demos have minor issues while running. Demo.WMS doesn't compile so must be unloaded to compile the rest of the project. It needs to be re-written with ASP.Net in order to work. I'll fix bugs in the library as I discover them, but likely not Demo.WMS